+++ HotBars/AmmoHotBar.cs
@@ -91,10 +91,7 @@ public static class AmmoSlotsHotBar
 
     internal static bool IsEnabled() => ammoSlotsHotBarEnabled.Value;
 
-    internal static bool IsShortcutDownWithItem(Slot slot) => slot.IsShortcutDown() && slot.Item != null;
+    internal static Slot GetSlotWithShortcutDown() => IsEnabled() ? hotBarSlots.FirstOrDefault(slot => slot.IsShortcutDownWithItem()) : null;
 
-    internal static Slot GetSlotWithShortcutDown() => IsEnabled() ? hotBarSlots.FirstOrDefault(IsShortcutDownWithItem) : null;
-
-    internal static IEnumerable<Slot> GetSlotsWithShortcutDown() => IsEnabled() ? hotBarSlots.Where(IsShortcutDownWithItem) : null;
-
+    internal static IEnumerable<Slot> GetSlotsWithShortcutDown() => IsEnabled() ? hotBarSlots.Where(slot => slot.IsShortcutDownWithItem()) : null;
 }